+++ packages/indexer/src/core/utils.ts
@@ -132,12 +132,14 @@ export function createLogger(
   });
 }
 
-function formatEVM(s: string) {
-  s = s.toLowerCase();
-  if (s.length === 42 && s.substring(0, 2) === '0x') {
+export function formatEVM(address: string) {
+  let s = address.toLowerCase();
+  if (s.substring(0, 2) === '0x') {
     s = s.substring(2);
-  } else if (s.length !== 40) {
-    throw new Error(`Not EVM address`);
+  }
+
+  if (s.length !== 40 && s.length !== 64) {
+    throw new Error(`Neither EVM nor interchain address: initial ${address}, modified and failed: ${s}`);
   }
 
   const digest = keccak256(toUtf8Bytes(s)).substring(2);
@@ -164,7 +166,7 @@ function formatEVM(s: string) {
 }
 
 export class Padded {
-  private s: string;
+  private s: string; // should start with 0x
 
   constructor(s: string) {
     if (s.length !== 66)
@@ -177,6 +179,8 @@ export class Padded {
   static fromEVM(s: string): Padded {
     if (s.length !== 42)
       throw new Error(`Input string length must be 42, got: ${s.length}`);
+    if (s.slice(0, 2) !== '0x')
+      throw new Error(`Input string length must start with '0x', got: ${s}`);
 
     return new Padded('0x' + '00'.repeat(12) + s.slice(2));
   }
@@ -188,6 +192,13 @@ export class Padded {
       return new Padded(s);
     }
     return s;
+  }
+
+  pretty() {
+    if (this.s.substring(0, 24) === '0x0000000000000000000000') {
+      return this.toEVMAddress();
+    }
+    return formatEVM(this.s)
   }
 
   toEVMAddress() {
@@ -266,4 +277,15 @@ export function shuffle<V>(array: V[]): V[] {
 
 export function onlyUnique<T>(value: T, index: number, self: T[]): boolean {
   return self.indexOf(value) === index;
+}
+
+
+export function randomString(length: number) {
+  let result           = '';
+  const characters       = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+  const charactersLength = characters.length;
+  for ( var i = 0; i < length; i++ ) {
+    result += characters.charAt(Math.floor(Math.random() * charactersLength));
+ }
+ return result;
 }