+++ packages/indexer/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- feat: exposing api metrics on different port
+- feat: blocks to tip metric
 - feat: add version endpoint
 - feat: Graphql
 - feat: RPC rate limiting